app/src/main/java/com/humzacov/oragniseme/AddReminder/Reminder.java
package com.humzacov.oragniseme.AddReminder;

public class Reminder {
    private String title;
    private String description;
    private String photo; //store link to photo, not photo itself
    private String recordings; //store link, not speech itself
    private String date;
    private String time;
    private String owner;
    private String privacySettings;
    private String location;
    private String shareCode;
    private String repeatSettings;

    public Reminder() {

    }

    public String getTitle() {
        return title;
    }

    public void setTitle(String title) {
        this.title = title;
    }

    public String getDescription() {
        return description;
    }

    public void setDescription(String description) {
        this.description = description;
    }

    public String getDate() {
        return date;
    }

    public void setDate(String date) {
        this.date = date;
    }

    public String getTime() {
        return time;
    }

    public void setTime(String time) {
        this.time = time;
    }

    public String getOwner() {
        return owner;
    }

    public void setOwner(String owner) {
        this.owner = owner;
    }

    public String getPrivacySettings() {
        return privacySettings;
    }

    public void setPrivacySettings(String privacySettings) {
        this.privacySettings = privacySettings;
    }

    public String getShareCode() {
        return shareCode;
    }

    public void setShareCode(String shareCode) {
        this.shareCode = shareCode;
    }

    public String getRepeatSettings() {
        return repeatSettings;
    }

    public void setRepeatSettings(String repeatSettings) {
        this.repeatSettings = repeatSettings;
    }

    public String getPhoto() {
        return photo;
    }

    public void setPhoto(String photo) {
        this.photo = photo;
    }

    public String getRecordings() {
        return recordings;
    }

    public void setRecordings(String recordings) {
        this.recordings = recordings;
    }

    public String getLocation() {
        return location;
    }

    public void setLocation(String location) {
        this.location = location;
    }
}